refactor: MAT-269 선수 교체 기능을 제공하는 조합 컴포넌트를 features/playerSubstitution으로 이관
