+++ src/features/playerSubstitution/components/PlayerGrid/index.ts
@@ -1,0 +1,1 @@
+export * from './PlayerGrid';
