+++ src/features/playerSubstitution/components/PlayerList/index.ts
@@ -1,0 +1,1 @@
+export * from './PlayerList';
