+++ src/features/playerSubstitution/components/SimplePlayerList/SimplePlayerList.stories.css.ts
