+++ src/features/playerSubstitution/index.ts
@@ -1,1 +1,3 @@
+export * from './components';
+export * from './PlayerSubstitutionPolicies';
 export * from './PlayerSubstitutionAdapter';
